+++ src/operation.h
@@ -24,6 +24,8 @@ void op_get_reg_imm(struct operation* op, uint8_t* d, uint8_t* K);
 
 void op_get_reg16_imm(struct operation* op, uint8_t* d, uint8_t* K);
 
+void op_get_reg_displacement(struct operation* op, uint8_t* d, uint8_t* q);
+
 void op_get_io_direct(struct operation* op, uint8_t* A, uint8_t* d);
 
 void op_get_rel_addr(struct operation* op, int16_t* K);